requirements.txt
# Web framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20

# Machine Learning
torch==2.8.0
torchvision==0.23.0
ultralytics==8.3.199

# Image Processing
opencv-python==4.12.0.88
Pillow==11.3.0
numpy==2.2.6

# Data Science
pandas==2.3.2
scikit-learn==1.7.2

# Utilities
requests==2.32.5
pydantic==2.11.8